+++ benchmarks/src/report.rs
@@ -47,7 +47,6 @@ pub fn write_markdown(
 - Recall@10: {:.4}\n\
 - MRR: {:.4}\n\
 - NDCG@10: {:.4}\n\
-- Association Success: {:.4}
 - Abstain precision: {:.4}\n\
 - Abstain recall: {:.4}\n\
 - Ingest ms: {}\n\
@@ -76,7 +75,6 @@ pub fn write_markdown(
         summary.recall_at_10,
         summary.mrr,
         summary.ndcg_at_10,
-        summary.association_success_rate,
         summary.abstain_precision,
         summary.abstain_recall,
         summary.ingest_ms,
