+++ llms-full.txt
@@ -645,6 +645,10 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 
 - impl WasmFramework
 
+### src/wasm_graph_rag.rs
+
+- impl WasmFramework
+
 ## README.md
 
 ### CLI workflow test (inject → probe → export → import)
@@ -4428,8 +4432,16 @@ impl WasmFramework {
     pub fn on_event(&self, callback: Function);
     pub fn inject_text(&self, id: String, text: String) -> Result<(), JsValue>;
     pub fn probe_text(&self, query: String, top_k: usize) -> Result<Array, JsValue>;
+}
+```
+
+## src/wasm_graph_rag.rs
+
+### impl WasmFramework
+
+```rust
+impl WasmFramework {
     pub fn probe_with_graph(&self, vector: &[u8], anchor_top_k: usize, max_hops: usize, min_assoc_strength: f32, similarity_weight: f32, graph_weight: f32, final_top_k: usize) -> Result<Array, JsValue>;
     pub fn probe_text_with_graph(&self, text: String, anchor_top_k: usize, max_hops: usize, min_assoc_strength: f32, similarity_weight: f32, graph_weight: f32, final_top_k: usize) -> Result<Array, JsValue>;
 }
 ```
-
