+++ src/lib.rs
@@ -226,6 +226,10 @@ pub mod prelude {
 pub mod wasm;
 #[cfg(target_arch = "wasm32")]
 mod wasm_ext;
+#[cfg(target_arch = "wasm32")]
+mod wasm_graph_rag;
 // Include wasm_ext for tests to run underlying data pattern tests on native
 #[cfg(all(test, not(target_arch = "wasm32")))]
 mod wasm_ext;
+#[cfg(all(test, not(target_arch = "wasm32")))]
+mod wasm_graph_rag;
